% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MWAD.R
\name{MWAD}
\alias{MWAD}
\title{Mean weighted absolute deviance (MWAD)}
\usage{
MWAD(x, weights = NULL, target = 1L)
}
\arguments{
\item{x}{A \code{numeric} vector for which the MWAD should be computed.}

\item{weights}{A \code{numeric} vector of the same length as  \code{x} with weights. Default is
NULL (no weights).}

\item{target}{A \code{numeric} value with the target from which the distance should be measured.}
}
\value{
A \code{numeric} with MWAD.
}
\description{
This function calculates MWAD for a given numeric vector and target.
}
\examples{
x <- 1:3
res <- secondPack::MWAD(x)
}
\seealso{
\code{\link[firstPack]{CalculateMWAD}}
}
